+++ fs/doc.go
@@ -1,5 +1,0 @@
-// Copyright 2021 Ross Light
-// SPDX-License-Identifier: ISC
-
-// Package fs provides a fallback for the io/fs package in Go 1.16.
-package fs